backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.3.0
pydantic-settings>=2.0.3
langchain>=0.0.195
python-docx>=0.8.11
python-pptx>=0.6.21
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2
qdrant-client>=1.4.0
sentence-transformers>=2.2.2
minio>=7.1.15
python-jose>=3.3.0
python-dotenv>=1.0.0
tenacity>=8.2.2
numpy>=1.24.3
pandas>=2.0.3
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.24.1
openai>=0.10.2 
